Added the prod-build gulp task
'prod-build' is a copy of the 'build' gulp task. With the exception of also minifying the end file. It ends up taking longer than the 'build' task but outputs a smaller file. During development the standard 'build' gulp task since it's quicker and we don't care too much about the size of the app.js file. For production the 'prod-build' should be ran, taking a bit longer but results in a smaller app and less for the user to download.