Standardized predicates.
Using first name and second name FOAF predicates instead of FOAF name, makes it easier to parse input and work with the values. Email is now in the form of mailto:name@gmail.com, this is the form it seems to be in all rdf files out there. Besides that I modified the edit profile page a bit, now it works almost as intended. We still can't really change the profile picture from there. There were some bugs as well in the code that would update your profile. All that is now fixed. ATM all seems to work as intended, I'll keep testing and trougleshoot as I go.