Backend ready.
We have the backend for removing nodes. Algorithm is fairly simple. Get the rdf file, parse for triples, add all triples except the one selected for removal into a new writer object, then serialize it and put it back to the originall adress.
Have to investigate if we can do patch requests or something along those lines, since putting and getting large files can be a bit daunting.