Further polishing of webId support.
Now the URI to store files is taken from the NIC('storage') triple that
should be present in any RDF file. This way we avoid having to deduce it
based on the URI (that is bug prone, and a terrible idea in general).
Besides that we also put ACL files correctly upon file creation. All
access controll is delegated to the ACL files, that means our app DOES
NOT decide whether a user is allowed to put a resource somewhere or not.
This is decided by the acl file at the respective uri. Makes everything
a lot more consistent. A problem is we have no more write access
function. Therefore we can't know beforehand if an user has access to
put a resource to a certain uri before he actually puts it. This
introduces some complications. I'm discussing it with the Solid team
at the moment. Besides that the feature where if you try to link / add
to a node where you have no access right it'll add / link to your home
node instead has been removed for the time being. Will be re-introduced
when a more appropriate front end is there and when we have a good write
access deduction funtion. Besides that we can remove links (NOT NODES!).
That will have to be integrated with a certain frontend as well.