Hooking up the delete functionality.
Combining the backend (rdf file / triple deletion) with the frontend
(the node popping out of the graph dynamically). Quite a bit of stuff to
decide on in terms of how the information will flow around. The delete
is not functional at the moment! Will be by the time of the next commit.