Adding images works, editing profile works.
Fixed the edit profile tab. The current setup has an issue. When
uploading a file we need to send a options request to the future
location of the file to get it's ACL location from there. The problem is
that the proxy doesn't proxy the options request correctly atm. So we
just send the options straight to the uri, ignoring the proxy. This
shouldn't be a problem, since OPTIONS shouldn't require any type of
authentification / authorization.
Also we have a logout request sent to the proxy when the user clicks log
out.