Polished the profile code.
Now the name, second name, email and profile pictures are fetched and
displayed in the nav-bar correctly. The code handling the update of the
profile has not been changed yet. Not sure if it works 100%. I did my
best not to change the info flow / logic much and rather fix small
errors here and there.