div overlay appears on 'info' icon click
Right now I added an 'info' icon on the top right, next to chat and
search, used solely for toggling the display of a div overlay. Clicking
on the icon displays a translucent div that covers the whole screen,
with some text inside to be replaced with the actual gesture indicator
images. Clicking anywhere on the div closes the overlay. This only
happens once and needs to be made so it occurs each time the icon is
clicked. Needs further styling/transitions etc.