Update AuthenticationConsent with new style structure
There are styles with the same settings in the typography style file
now. I think this is OK as there is no guarantee that they will remain
the same in the future, but they will be used in different components.

The two in question at the moment are `cardSecondaryText` and
`sectionHeader`.

They are technically used in different contexts, and this means that
diverging them in the future is not necessary.