Actually make text size names consistent
Spacing uses an enum, where keys cannot begin with an integer if we
want to use dot notation to access the values. We are using XXS and XXL,
which is not consistent with 3XS, 4XL etc. but I think the trade-off is
fine as most of our values should fall within the XXS - XXL range. I also believe that the slight inconsistency is worth the trade-off.