Incorporate small render method in main render
I think we originally had this pattern of smaller render methods to keep
the main render method cleaner, but I'm not sure if it makes it harder to
understand or not. I have removed one of these smaller methods and have
added a TODO to replace the last one with issuerCard. I think in general,
if we want to have a cleaner main render method, we should use new
components and files.