Fix error with selecting claims on consent
I broke the sharing of claims during the  restructure of the consent screen
into children components. The handleClaimSelect function was passed down
via props and the binding of `this` was called from the child component
rather than being bound to the class. The class method has been changed to
an arrow function to bind `this` to the class.