fix(ios/AppDelegate.m): remove unnecessary bits about dev build
iOS devices access the dev server over the network, and the device needs
to be on the same network for this to work. And the network needs to
allow it of course, so this was non-functional in Betahaus. This is
probably the only reference to Betahaus in our git logs; this is their
legacy.
Anyway, there's an xcode build phase script
(node_modules/react-native/scripts/react-native-xcode.sh) which
"guesses" the build machine's IP and writes it to a text file inside the
.ipa, which is then loaded by the react bridge.

Closes #1396